requirements.txt
    flask
    pinecone-client
    sentence-transformers
    python-dotenv
    gunicorn
    ```
